requirements.txt
aiohttp==3.11.7
anyio==4.6.2.post1
appnope==0.1.4
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==5.1.1
executing==2.1.0
faiss-cpu==1.9.0.post1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langsmith==0.1.145
marshmallow==3.23.1
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.56.2
orjson==3.10.12
packaging==24.2
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0